+++ lib/dexter/processor.rb
@@ -23,6 +23,8 @@ module Dexter
           JsonLogParser.new(logfile, @collector)
         elsif options[:input_format] == "sql"
           SqlLogParser.new(logfile, @collector)
+        elsif options[:input_format] == "supabase"
+          SupabaseLogParser.new(logfile, @collector)
         else
           StderrLogParser.new(logfile, @collector)
         end